// src/types/ShortenUrl.ts
// Generated by https://quicktype.io

export interface ShortenURL {
    ok: boolean;
    result: Result;
}

export interface Result {
    code: string;
    short_link: string;
    full_short_link: string;
    short_link2: string;
    full_short_link2: string;
    short_link3: string;
    full_short_link3: string;
    share_link: string;
    full_share_link: string;
    original_link: string;
}

export type LocalShortenUrl = {
    originalUrl: string;
	shortenUrl: string;
	fullShortenUrl: string;
    id: string;
};